==> scripts/inflammatory_meta/3b_inflammatory_DE_replication.Rmd <==
---
title: "Inflammatory DE testing"
author: "NSH"
output: html_notebook
---

```{r setup}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r, message=FALSE}
if (!require(BiocManager)) install.packages("BiocManager")
if (!require(DESeq2)) BiocManager::install("DESeq2")
if (!require(apeglm)) BiocManager::install("apeglm")
if (!require(qvalue)) BiocManager::install("qvalue")

#qvalue_truncp export is broken in the current version, manually add the fx here, (c) by StoreyLab (https://github.com/StoreyLab/qvalue/blob/9b3f9a8af4dc8b680a6914d1fcc106e041b7c7a0/R/qvalue_trunc.R):
qvalue_truncp <- function(p, fdr.level = NULL, pfdr = FALSE, lfdr.out = TRUE, pi0 = NULL, ...) {
  # Argument checks
  p_in <- qvals_out <- lfdr_out <- p
  rm_na <- !is.na(p)
  p <- p[rm_na]
  if (min(p) < 0 || max(p) > 1) {
    stop("p-values not in valid range [0, 1].")
  } else if (!is.null(fdr.level) && (fdr.level <= 0 || fdr.level > 1)) {
    stop("'fdr.level' must be in (0, 1].")
  }
  p <- p / max(p)
  # Calculate pi0 estimate
  if (is.null(pi0)) {
    pi0s <- pi0est(p, ...)
  } else {
    if (pi0 > 0 && pi0 <= 1)  {
      pi0s = list()
      pi0s$pi0 = pi0
    } else {
      stop("pi0 is not (0,1]")
    }
  }
  
  # Calculate q-value estimates
  m <- length(p)
  i <- m:1L
  o <- order(p, decreasing = TRUE)
  ro <- order(o)
  if (pfdr) {
    qvals <- pi0s$pi0 * pmin(1, cummin(p[o] * m / (i * (1 - (1 - p[o]) ^ m))))[ro]
  } else {
    qvals <- pi0s$pi0 * pmin(1, cummin(p[o] * m /i ))[ro]
  }
  qvals_out[rm_na] <- qvals
  # Calculate local FDR estimates
  if (lfdr.out) {
    lfdr <- lfdr(p = p, pi0 = pi0s$pi0, ...)
    lfdr_out[rm_na] <- lfdr
  } else {
    lfdr_out <- NULL
  }
  
  # Return results
  if (!is.null(fdr.level)) {
    retval <- list(call = match.call(), pi0 = pi0s$pi0, qvalues = qvals_out,
                   pvalues = p_in, lfdr = lfdr_out, fdr.level = fdr.level,
                   significant = (qvals <= fdr.level),
                   pi0.lambda = pi0s$pi0.lambda, lambda = pi0s$lambda,
                   pi0.smooth = pi0s$pi0.smooth)
  } else {
    retval <- list(call = match.call(), pi0 = pi0s$pi0, qvalues = qvals_out,
                   pvalues = p_in, lfdr = lfdr_out, pi0.lambda = pi0s$pi0.lambda,
                   lambda = pi0s$lambda, pi0.smooth = pi0s$pi0.smooth)
  }
  class(retval) <- "qvalue"
  return(retval)
}
  


source("scripts/utils/utils.R")
source("scripts/utils/figure_builder.R")
```

```{r}
res_list <- readRDS("data/processed/de_seq_res_list.rds")
```

# Pi1-replicability

Here, we test the replicability of individual differential expression results using the pi1 statistic.

```{r}
replication_res <- lapply(seq_along(res_list), function(i) {
  res_outer <- res_list[[i]]
  name_outer <- names(res_list)[[i]]
  
  sig_events <- res_outer %>%
    as.data.frame() %>%
    rownames_to_column("symbol") %>%
    filter(!is.na(padj),
           padj < 0.05,
           abs(log2FoldChange) >= 0.5) %>%
    pull(symbol)
  
  lapply(seq_along(res_list), function(i) {
    res_inner <- res_list[[i]]
    name_inner <- names(res_list)[[i]]
    
    hits <- res_inner %>%
      as.data.frame() %>%
      rownames_to_column("symbol") %>%
      filter(!is.na(pvalue),
             is.finite(pvalue),
             symbol %in% sig_events) %>%
      pull(pvalue, symbol)
    
    tryCatch(
      {
        q <- qvalue_truncp(hits, fdr.level = 0.05)
        pi1 <- 1-q$pi0
        return (data.frame(events_from = name_outer,
                           hits_from = name_inner,
                           pi1 = pi1))
      },
      error = function(e) {
        return (data.frame(events_from = name_outer,
                           hits_from = name_inner,
                           pi1 = NA))
      }
    )

  }) %>%
    Reduce(function(df1, df2) rbind(df1, df2), .)
}) %>%
  Reduce(function(df1, df2) rbind(df1, df2), .)
```


```{r}
replication_res %>%
  filter(events_from != hits_from) %>%
  mutate(events_spec = str_remove(events_from, "^.*_"),
         hits_spec = str_remove(hits_from, "^.*_")) %>%
  ggplot(aes(x = events_from, y = pi1, color = hits_spec)) +
    geom_boxplot(aes(color = NULL)) +
    geom_jitter() +
    theme(axis.text.x = element_text(angle = 45,
                                     vjust = 1,
                                     hjust = 1)) +
    facet_wrap(~events_spec,
               scales = "free_x")
```


# Correlation HM

Here, the correlation between the LFCs of differentially expressed genes are calculated.

```{r}
lfc_res <- lapply(seq_along(res_list), function(i) {
  res_outer <- res_list[[i]]
  name_outer <- names(res_list)[[i]]
  
  lfc_outer <- res_outer %>%
    as.data.frame() %>%
    rownames_to_column("symbol") %>%
    filter(!is.na(padj),
           padj < 0.05,
           abs(log2FoldChange) >= 0.5) %>%
    transmute(symbol, lfc_outer = log2FoldChange)
  
  lapply(seq_along(res_list), function(i) {
    res_inner <- res_list[[i]]
    name_inner <- names(res_list)[[i]]
    
    lfc_inner <- res_inner %>%
      as.data.frame() %>%
      rownames_to_column("symbol") %>%
      filter(!is.na(padj),
             padj < 0.05,
             abs(log2FoldChange) >= 0.5) %>%
      transmute(symbol, lfc_inner = log2FoldChange)
    
    io <- inner_join(lfc_inner, lfc_outer, by = "symbol")
    
    if (!nrow(io) %in% c(0, 1, 2)){
      cor_io <- cor.test(io$lfc_inner, io$lfc_outer)
      
      return (tibble(
        name_outer = name_outer,
        name_inner = name_inner,
        cor = cor_io$estimate,
        p = cor_io$p.value,
        size_outer = nrow(lfc_outer),
        size_inner = nrow(lfc_inner),
        overlap = nrow(io),
        rel_overlap = overlap/size_outer,
        method = cor_io$method
        ))
    } else {
      return (tibble(
        name_outer = name_outer,
        name_inner = name_inner,
        cor = NA,
        p = NA,
        size_outer = nrow(lfc_outer),
        size_inner = nrow(lfc_inner),
        overlap = nrow(io),
        rel_overlap = overlap/size_outer,
        method = NA
        ))
    }
    

  }) %>%
    Reduce(function(df1, df2) rbind(df1, df2), .)
}) %>%
  Reduce(function(df1, df2) rbind(df1, df2), .) %>%
  group_by(name_outer) %>%
  mutate(padj = p.adjust(p, method = "BH"), .after = p) %>%
  ungroup()
```

## Visualize as heatmap

```{r}
lfc_cor_mat <- lfc_res %>%
  pivot_wider(id_cols = name_outer,
              names_from = name_inner,
              values_from = cor) %>%
  column_to_rownames("name_outer")
lfc_size_mat <- lfc_res %>%
  pivot_wider(id_cols = name_outer,
              names_from = name_inner,
              values_from = overlap) %>%
  column_to_rownames("name_outer")
Heatmap(lfc_cor_mat,
        cell_fun = function(j, i, x, y, width, height, fill) {
          grid.text(sprintf("%.0f", lfc_size_mat[i, j]), x, y, gp = gpar(fontsize = 10))
          }
        )
```

## Visualize as boxplot/jitter

```{r}
ggplot(lfc_res %>%
         filter(name_outer != name_inner) %>%
         mutate(padj = ifelse(padj < 0.05, padj, NA)),
       aes(x = name_outer, y = cor, color = -log10(padj), size = rel_overlap)) +
  geom_boxplot(aes(color = NULL, size = NULL), outlier.size = NULL) +
  geom_jitter()
```

```{r}
save(lfc_res, replication_res, file = "data/processed/DE_testing_replicability.rda")
```
