==> scripts/inflammatory_meta/5a_inflammatory_fisher_profiling.Rmd <==
---
title: "Profiling of inflammatory fisher meta signature"
author: "NSH"
output: html_notebook
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

```{r message=FALSE, include=FALSE}
if (!require(tidyverse)) install.packages("tidyverse")
if (!require(magrittr)) install.packages("magrittr")
if (!require(ggplot2)) install.packages("ggplot2")
if (!require(RColorBrewer)) install.packages("RColorBrewer")

if (!require(BiocManager)) install.packages("BiocManager")
if (!require(DESeq2)) BiocManager::install("DESeq2")
if (!require(vsn)) BiocManager::install("vsn")
if (!require(ComplexHeatmap)) BiocManager::install("ComplexHeatmap")
if (!require(circlize)) BiocManager::install("circlize")
if (!require(sva)) BiocManager::install("sva")
if (!require(edgeR)) BiocManager::install("edgeR")

# import color palettes
source("scripts/utils/utils.R")
source("scripts/utils/figure_builder.R")
```

```{r include=FALSE}
cross_se <- readRDS("data/processed/cross_se.rds")
cross_se_inflammatory <- cross_se[, cross_se$inflammatory_analysis]

p_df <- readRDS("data/processed/fisher_p_df.rds")
lfc_df <- readRDS("data/processed/fisher_lfc_df.rds")
p_lfc_df <- readRDS("data/processed/fisher_p_lfc_df.rds")
up_genes <- readRDS("data/processed/fisher_up_genes.rds")
dn_genes <- readRDS("data/processed/fisher_dn_genes.rds")
core_genes <- readRDS("data/processed/up_fisher_human.rds")
```

# Gene rank plot

```{r, fig.height=5, fig.width=20}
plt <- p_lfc_df %>%
  arrange(fisher_adjusted) %>%
  mutate(rank = 1:length(fisher_adjusted))

ggplot (plt, aes (x = rank, y = -log10(fisher_adjusted))) +
  geom_point()
ggplot (plt %>% filter(mean_lfc > 0), aes (x = rank, y = -log10(fisher_adjusted))) +
  geom_point()
ggplot (plt %>% filter(mean_lfc < 0), aes (x = rank, y = -log10(fisher_adjusted))) +
  geom_point()
ggplot (plt %>% filter(mean_lfc > 0.5), aes (x = rank, y = -log10(fisher_adjusted))) +
  geom_point()
ggplot (plt %>% filter(mean_lfc < -0.5), aes (x = rank, y = -log10(fisher_adjusted))) +
  geom_point()


p_lfc_df %>%
  arrange (fisher_adjusted) %>%
  transmute (rank = seq_len(nrow(.)),
             p_fisher_log = -log10(fisher_adjusted)) %>%
  ggplot (mapping = aes (x = rank, y = p_fisher_log)) +
    geom_point() +
    geom_vline(mapping = aes(xintercept = 500)) +
    geom_hline(yintercept = -log10(0.05))# +
    #coord_trans(x = "log10") +
    #scale_x_continuous(breaks = c(0, 500, 2500, 5000, 10000))
```

# Signature Volcano

```{r}
volcano_stats <- lfc_df %>%
  pivot_longer(
    cols = -"symbol",
    names_to = "study",
    values_to = "lfc"
  ) %>%
  full_join(p_df %>%
              dplyr::select(symbol, fisher_adjusted),
            by = "symbol") %>%
  group_by(symbol) %>%
  transmute(pval = fisher_adjusted,
            lfc = mean(lfc)) %>%
  ungroup() %>%
  distinct()

saveRDS(volcano_stats, "data/processed/data_fig_4b.rds")

p_lfc_rank_stats <- p_lfc_df %>%
  arrange (fisher_adjusted) %>%
  mutate(
    fisher_rank = seq_len(nrow(.)),
    fisher_adjusted_fixed = ifelse(
      fisher_adjusted == 0,
      min(fisher_adjusted[fisher_adjusted != 0])*1e-1,
      fisher_adjusted
      ),
    p_fisher_log = -log10(fisher_adjusted),
    p_fisher_log_fixed = -log10(fisher_adjusted_fixed))

saveRDS(p_lfc_rank_stats, "data/processed/p_lfc_rank_stats.rds")
```


```{r}
# fig_4b <- build_fig_4b()
# fig_4b
```


# LFC heatmaps

```{r}
heatmap_matrix <- lfc_df %>%
  filter (symbol %in% c(up_genes, dn_genes)) %>%
  #filter(symbol %in% sig_genes) %>%
  pivot_longer(!symbol, names_to = "study", values_to = "lfc") %>%
  group_by(symbol) %>%
  filter(!any(is.na(lfc))) %>%
  mutate(mean = mean(lfc)) %>%
  ungroup() %>%
  arrange(mean) %>%
  pivot_wider(id_cols = c(study), names_from = symbol, values_from = lfc) %>%
  mutate(study = str_remove(study, "lfc__")) %>%
  mutate(study = str_remove(study, "_vs_HC")) %>%
  column_to_rownames("study") %>%
  as.matrix()

heatmap_matrix_scaled <- lfc_df %>%
  filter (symbol %in% c(up_genes, dn_genes)) %>%
  pivot_longer(!symbol, names_to = "study", values_to = "lfc") %>%
  group_by(study) %>%
  mutate(lfc_scaled = scale(lfc)) %>%
  group_by(symbol) %>%
  filter(!any(is.na(lfc_scaled))) %>%
  mutate(mean = mean(lfc_scaled)) %>%
  ungroup() %>%
  arrange(mean) %>%
  pivot_wider(id_cols = c(study), names_from = symbol, values_from = lfc_scaled) %>%
  mutate(study = str_remove(study, "lfc__")) %>%
  mutate(study = str_remove(study, "_vs_HC")) %>%
  column_to_rownames("study") %>%
  as.matrix()

saveRDS(heatmap_matrix, "data/processed/data_fig_4c_1.rds")
saveRDS(heatmap_matrix_scaled, "data/processed/data_fig_4c_2.rds")
```

```{r}
# no row scaling
fig_3c_1 <- build_fig_3c_1()
# row scaling
fig_3c_2 <- build_fig_3c_2()

fig_3c_1
fig_3c_2
```

# LFC surfaceome heatmap

Filter for surfaceome using https://wlab.ethz.ch/surfaceome/ Cite: Bausch-Fluck et al. 2018 PNAS


Heatmap of filtered genes: LFC of gene per study

```{r}
surfaceome_data <- read_csv("data/metadata/table_S3_surfaceome.csv")

plt_surfaceome <- heatmap_matrix[ , colnames(heatmap_matrix) %in% surfaceome_data$`UniProt gene`]

saveRDS(plt_surfaceome, "data/processed/data_fig_6a.rds")
# biolegend_ab <- read.delim("./tables/Human Abs gene ID.txt")
# inflammatory_surfaceome_biolegend <- merge(inflammatory_surfaceome, biolegend_ab[,c("Specificity", "Cat_Num")], by.x="symbol", by.y = "Specificity", all.x=T)
# inflammatory_surfaceome_biolegend <- inflammatory_surfaceome_biolegend[!duplicated(inflammatory_surfaceome_biolegend$symbol),]
#  inflammatory_surfaceome_biolegend[order(inflammatory_surfaceome_biolegend$fisher_adjusted),]


```


```{r}
fig_6a <- build_fig_6a()
fig_6a

# colfun <- generate_rd_bu_colfun(-2, 2)
# 
# plt_surfaceome_scaled <- plt_surfaceome %>%
#  t() %>%
#  scale() %>%
#  t()

# pdf("figures/delete/conserved_surfaceome.pdf",
#     width = 10,
#     height = 10*9/20)

# dev.off()
```


# Expression heatmaps

```{r}
series <- factor(colData(cross_se_inflammatory)[["series"]])
expr <- as.matrix(assay(cross_se_inflammatory, "counts"))
expr_corrected <- ComBat_seq(expr, series)

expr_corrected_norm <- calcNormFactors(DGEList(expr_corrected), method = "TMM")
lcpm <- edgeR::cpm(expr_corrected_norm, log = TRUE, normalized.lib.sizes = TRUE)
lcpm_core <- lcpm[rownames(lcpm) %in% core_genes, ]
lcpm_scaled <- lcpm_core %>%
  t() %>%
  scale() %>%
  t()
```


```{r}
annotation <- cross_se_inflammatory %>%
  colData() %>%
  as.data.frame() %>%
  rownames_to_column("sample_id") %>%
  transmute(sample_id,
            condition = ifelse(condition == "HC", "HC", "INFL"),
            species) %>%
  filter(sample_id %in% colnames(lcpm_scaled)) %>%
  arrange(match(sample_id, colnames(lcpm_scaled)))

saveRDS(lcpm_scaled, "data/processed/data_1_fig_4d_1.rds")
saveRDS(annotation, "data/processed/data_2_fig_4d_1.rds")
```

Generate data that is sorted and does not need clustering.

```{r}
lcpm_sorted <- lcpm_scaled %>%
  as.data.frame() %>%
  rownames_to_column("symbol") %>%
  pivot_longer(cols = -"symbol",
               names_to = "sample_id",
               values_to = "count") %>%
  left_join(annotation, by = "sample_id") %>%
  arrange(condition, species)

lcpm_sorted_rank <- lcpm_sorted %>%
  filter(condition == "INFL") %>%
  group_by(condition, symbol) %>%
  summarise(mean_infl = mean (count)) %>%
  arrange(desc(mean_infl))

lcpm_sorted_wide <- lcpm_sorted %>%
  pivot_wider(id_cols = c("symbol"),
              names_from = "sample_id",
              values_from = "count") %>%
  column_to_rownames("symbol")


annotation_sorted <- annotation %>%
  arrange(condition, species)

saveRDS(lcpm_sorted_rank, "data/processed/data_1_fig_4d_2.rds")
saveRDS(lcpm_sorted_wide, "data/processed/data_2_fig_4d_2.rds")
saveRDS(annotation_sorted, "data/processed/data_3_fig_4d_2.rds")
```

Build the figures.
```{r}
fig_3d_1 <- build_fig_3d_1()
#fig_3d_2 <- build_fig_3d_2()

fig_3d_1
#fig_3d_2
```

```{r}
#fig_3c_2 <- build_fig_3c_2()
#pdf("figures/delete/hc_infl_expr_hm.pdf",
#    width = 10,
#    height = 10*9/16)
draw_fig_3d_2()
#dev.off()
```
